+++ backend/src/validators/invoices.js
@@ -1,40 +1,63 @@
 const { Schema, getResponse, getUpdateResponse, getAddResponse } = require("./common_schema.js");
 const S = require("fluent-json-schema");
 
-const body = S.object()
-  .prop("id", Schema.Id)
+const getAll = {
+  params: Schema.IdParam,
+  response: getResponse(
+    S.array().items(
+      S.object()
+        .prop("id", S.number())
+        .prop("received_date", S.string())
+        .prop("invoice_date", S.string())
+        .prop("due_date", S.string())
+        .prop("billing_period", S.string())
+        .prop("fiscal", S.string())
+        .prop("invoice_total", S.number())
+        .prop("invoice_number", S.string())
+        .prop("invoice_total", S.number())
+        .prop("is_gl", S.boolean())
+        .prop("notes", S.string())
+    )
+  ),
+};
+
+const getOne = {
+  params: Schema.IdParam,
+  response: getResponse(
+    S.object()
+      .prop("id", S.number())
+      .prop("received_date", S.string())
+      .prop("invoice_date", S.string())
+      .prop("due_date", S.string())
+      .prop("billing_period", S.string())
+      .prop("fiscal", Schema.Picker)
+      .prop("invoice_number", S.string())
+      .prop("invoice_total", S.number())
+      .prop("is_gl", S.boolean())
+      .prop("notes", S.string())
+  ),
+};
+
+const addUpdateBody = S.object()
   .prop("received_date", Schema.RequiredDate)
   .prop("invoice_date", Schema.Date)
   .prop("due_date", Schema.Date)
   .prop("billing_period", Schema.ShortString)
-  .prop("fiscal", Schema.ShortString)
-  .prop("invoice_total", Schema.Money)
+  .prop("fiscal", Schema.Id)
   .prop("invoice_number", Schema.ShortString)
   .prop("invoice_total", Schema.Money)
   .prop("is_gl", S.boolean())
   .prop("notes", S.string());
 
-const singleBody = body.without(["fiscal", "invoice_total"]).prop("fiscal", Schema.Picker);
-
-const getAll = {
-  params: Schema.IdParam,
-  response: getResponse(S.array().items(body)),
-};
-
-const getOne = {
-  params: Schema.IdParam,
-  response: getResponse(singleBody),
-};
-
 const updateOne = {
   params: Schema.IdParam,
-  body: body.without(["id", "fiscal", "invoice_total"]).prop("fiscal", Schema.Id),
+  body: addUpdateBody,
   response: getUpdateResponse(),
 };
 
 const addOne = {
   params: Schema.IdParam,
-  body: body.without(["id", "fiscal", "invoice_total"]).prop("fiscal", Schema.Id),
+  body: addUpdateBody,
   response: getAddResponse(),
 };