+++ backend/src/validators/report.js
@@ -3,7 +3,7 @@ const S = require("fluent-json-schema");
 
 const getOne = {
   params: Schema.IdParam,
-  response: getResponse(S.object().prop("projectId", Schema.Id)),
+  response: getResponse(S.object().prop("projectId", S.number())),
 };
 
 module.exports = {