+++ backend/src/validators/resources.js
@@ -1,46 +1,50 @@
 const { Schema, getResponse, getAddResponse, getUpdateResponse } = require("./common_schema.js");
 const S = require("fluent-json-schema");
 
-const multiBody = S.object()
-  .prop("id", Schema.Id)
-  .prop("last_name", Schema.ShortString)
-  .prop("first_name", Schema.ShortString)
-  .prop("supplier", Schema.ShortString)
-  .prop("subcontractor", Schema.ShortString)
-  .prop("created_date", Schema.ShortString);
-
-const singleBody = S.object()
-  .prop("id", Schema.Id)
-  .prop("created_date", Schema.Date)
-  .prop("created_date_formatted", Schema.ShortString)
-  .prop("resource_first_name", Schema.ShortString)
-  .prop("resource_last_name", Schema.ShortString)
-  .prop("subcontractor_id", Schema.Picker)
-  .prop("supplier_id", Schema.Picker);
-
-const requestBody = singleBody
-  .without(["created_date_formatted", "subcontractor_id", "supplier_id"])
-  .minProperties(1)
-  .prop("subcontractor_id", Schema.Id)
-  .prop("supplier_id", Schema.Id);
-
 const getAll = {
-  response: getResponse(S.array().items(multiBody)),
+  response: getResponse(
+    S.array().items(
+      S.object()
+        .prop("id", S.number())
+        .prop("last_name", S.string())
+        .prop("first_name", S.string())
+        .prop("supplier", S.string())
+        .prop("subcontractor", S.string())
+        .prop("created_date", S.string())
+    )
+  ),
 };
 
 const getOne = {
   params: Schema.IdParam,
-  response: getResponse(singleBody),
+  response: getResponse(
+    S.object()
+      .prop("id", S.number())
+      .prop("created_date", S.string())
+      .prop("created_date_formatted", S.string())
+      .prop("resource_first_name", S.string())
+      .prop("resource_last_name", S.string())
+      .prop("subcontractor_id", Schema.Picker)
+      .prop("supplier_id", Schema.Picker)
+  ),
 };
+
+const addUpdateBody = S.object()
+  .prop("created_date", Schema.Date)
+  .prop("created_date_formatted", Schema.ShortString)
+  .prop("resource_first_name", Schema.ShortString)
+  .prop("resource_last_name", Schema.ShortString)
+  .prop("subcontractor_id", Schema.Id)
+  .prop("supplier_id", Schema.Id);
 
 const updateOne = {
   params: Schema.IdParam,
-  body: requestBody,
+  body: addUpdateBody,
   response: getUpdateResponse(),
 };
 
 const addOne = {
-  body: requestBody,
+  body: addUpdateBody,
   response: getAddResponse(),
 };