+++ backend/src/validators/suppliers.js
@@ -1,15 +1,46 @@
 const { Schema, getResponse, getAddResponse, getUpdateResponse } = require("./common_schema.js");
 const S = require("fluent-json-schema");
 
-const multiBody = S.object()
-  .prop("id", Schema.Id)
-  .prop("supplier_number", Schema.Id)
-  .prop("signing_authority", Schema.ShortString)
-  .prop("financial_contact", Schema.ShortString)
-  .prop("province", Schema.ShortString);
+const getAll = {
+  response: getResponse(
+    S.array().items(
+      S.object()
+        .prop("id", S.number())
+        .prop("supplier_number", S.number())
+        .prop("signing_authority", S.string())
+        .prop("financial_contact", S.string())
+        .prop("province", S.string())
+    )
+  ),
+};
 
-const singleBody = S.object()
-  .prop("id", Schema.Id)
+const getOne = {
+  params: Schema.IdParam,
+  response: getResponse(
+    S.object()
+      .prop("id", S.number())
+      .prop("address", S.string())
+      .prop("city", S.string())
+      .prop("country", S.string())
+      .prop("fax", S.string())
+      .prop("email", S.string())
+      .prop("financial_contact_email", S.string())
+      .prop("financial_contact_name", S.string())
+      .prop("financial_contact_phone", S.string())
+      .prop("phone", S.string())
+      .prop("postal_code", S.string())
+      .prop("province", S.string())
+      .prop("signing_authority_name", S.string())
+      .prop("signing_authority_title", S.string())
+      .prop("site_number", S.string())
+      .prop("supplier_legal_name", S.string())
+      .prop("supplier_name", S.string())
+      .prop("supplier_number", S.number())
+      .prop("website", S.string())
+  ),
+};
+
+const addUpdateBody = S.object()
   .prop("address", Schema.ShortString)
   .prop("city", Schema.ShortString)
   .prop("country", Schema.ShortString)
@@ -29,23 +60,14 @@ const singleBody = S.object()
   .prop("supplier_number", Schema.Id)
   .prop("website", Schema.Uri);
 
-const getAll = {
-  response: getResponse(S.array().items(multiBody)),
-};
-
-const getOne = {
-  params: Schema.IdParam,
-  response: getResponse(singleBody),
-};
-
 const updateOne = {
   params: Schema.IdParam,
-  body: singleBody.without(["id"]).minProperties(1),
+  body: addUpdateBody,
   response: getUpdateResponse(),
 };
 
 const addOne = {
-  body: singleBody.without(["id"]).minProperties(1),
+  body: addUpdateBody,
   response: getAddResponse(),
 };