+++ backend/src/validators/users.js
@@ -1,38 +1,49 @@
 const { Schema, getResponse, getAddResponse, getUpdateResponse } = require("./common_schema.js");
 const S = require("fluent-json-schema");
 
-const baseBody = S.object()
-  .prop("id", Schema.Id)
-  .prop("name", Schema.ShortString)
-  .prop("email", Schema.RequiredEmail)
-  .prop("role_id", Schema.Id);
-
 const getAll = {
   response: getResponse(
-    S.array().items(baseBody.without(["role_id"]).prop("user_role", Schema.ShortString))
+    S.array().items(
+      S.object()
+        .prop("id", S.number())
+        .prop("name", S.string())
+        .prop("email", S.string())
+        .prop("user_role", S.string())
+    )
   ),
 };
 
 const getOne = {
   params: Schema.IdParam,
-  response: getResponse(baseBody.prop("role_id", Schema.Picker)),
+  response: getResponse(
+    S.object()
+      .prop("id", S.number())
+      .prop("name", S.string())
+      .prop("email", S.string())
+      .prop("role_id", Schema.Picker)
+  ),
 };
+
+const addUpdateBody = S.object()
+  .prop("name", Schema.ShortString)
+  .prop("email", Schema.RequiredEmail)
+  .prop("role_id", Schema.Id);
 
 const updateOne = {
   params: Schema.IdParam,
-  body: baseBody.without(["id"]).minProperties(1),
+  body: addUpdateBody,
   response: getUpdateResponse(),
 };
 
 const addOne = {
   headers: S.object().prop("Authorization", S.string()),
-  body: baseBody.without(["id"]).required(["email", "name", "role_id"]),
+  body: addUpdateBody.required(["email", "name", "role_id"]),
   response: getAddResponse(),
 };
 
 const getByEmail = {
   headers: S.object().prop("Authorization", S.string()),
-  body: S.object().prop("email", Schema.Email).required(["email"]),
+  body: S.object().prop("email", S.string()).required(["email"]),
   response: getAddResponse(),
 };